SenseCom v1.3.0, SGConnect v1.5.1, SGCoreCs v1.5.1
SGConnect now stops connecting to devices that are paired, but turned off after 2 attempts. These connections can be retried through a function call. Pushed a SenseCom update that allows for this via a button, as well as an update the SGCore C# API to call this function through Interopservices. SGCoreCpp is awaiting a major overhaul. It's update will follow later.

SenseCom binaries now contained in zip folders for easier transfer.