Major Repository overhaul for v2.102 binaries and Nova 2.0
Removes C# Binaries from repository. Updates C++ binaries to a new v2.0 back-end, which introduces breaking changes to haptics and calibration (for the better). Also updates SenseCom to v1.6.1.
Removes doc folder for now, and example projects require a major refactor to cmake files to be easy to use. But one can use this to work with Nova 2.0,